+++ src/main/java/net/oneandone/maven/plugins/dockerbuild/Build.java
@@ -164,7 +164,6 @@ public class Build extends Base {
         a.addArtifacts(context, world.file(project.getBuild().getDirectory()), artifactName);
         a.addFiles(srcdirs(), buildDirectory(), fileFilter, project, session);
         a.addPom(project);
-        a.addProperty(project);
         a.addExplicit(arguments);
         actuals = a.result();
         try (InputStream tarSrc = context.tar().newInputStream()) {